aware-microservices-main/aware-microservices-main/requirements.txt
aniso8601==10.0.0
attrs==25.1.0
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
dnspython==2.7.0
docker==7.1.0
exceptiongroup==1.2.2
Flask==3.1.0
flask-restx==1.3.0
gunicorn==23.0.0
idna==3.10
importlib_resources==6.5.2
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.5
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
MarkupSafe==3.0.2
packaging==24.2
pika==1.3.2
pluggy==1.5.0
pymongo==4.10.1
pytest==8.3.4
python-dotenv==1.0.1
pytz==2025.1
referencing==0.36.2
requests==2.32.3
rpds-py==0.22.3
tomli==2.2.1
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.1.3
